feat: Refactor user registration and login methods, update role management
