+++ HealthCareAB_v1/HealthCareAB_v1/Controllers/AuthController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Security.Claims;
 using HealthCareAB_v1.Constants;
 using HealthCareAB_v1.DTOs;
+using HealthCareAB_v1.Exceptions;
 using HealthCareAB_v1.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -22,7 +23,7 @@ namespace HealthCareAB_v1.Controllers
         /// <summary>
         /// Registers a new user with default User role.
         /// </summary>
-        [HttpPost("register")]
+        [HttpPost("register-patient")]
         [ProducesResponseType(typeof(AuthResponseDto), StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
         public async Task<IActionResult> Register([FromBody] RegisterDto request)
@@ -30,7 +31,7 @@ namespace HealthCareAB_v1.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var result = await _authService.RegisterAsync(request);
+            var result = await _authService.RegisterPatientAsync(request);
 
             if (!result.Success)
             {
@@ -46,6 +47,98 @@ namespace HealthCareAB_v1.Controllers
                     roles = result.Roles,
                 }
             );
+        }
+
+        [HttpPost("login-patient")]
+        [ProducesResponseType(typeof(AuthResponseDto), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<IActionResult> Login([FromBody] LoginDto request)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            try
+            {
+                var (result, token) = await _authService.LoginPatientAsync(
+                    request.Identifier,
+                    request.Password
+                );
+
+                if (!result.Success || string.IsNullOrEmpty(token))
+                {
+                    if (result.IsLockedOut)
+                    {
+                        return StatusCode(
+                            StatusCodes.Status403Forbidden,
+                            new { message = result.Message }
+                        );
+                    }
+                    return Unauthorized(new { message = result.Message });
+                }
+
+                var cookieOptions = _authService.GetJwtCookieOptions();
+                HttpContext.Response.Cookies.Append(CookieNames.Jwt, token, cookieOptions);
+
+                return Ok(
+                    new
+                    {
+                        message = result.Message,
+                        loggedInUser = $"{result.FirstName} {result.LastName}",
+                        roles = result.Roles,
+                    }
+                );
+            }
+            catch (JwtTokenGenerationException ex)
+            {
+                return StatusCode(500, new { message = ex.Message });
+            }
+        }
+
+        [HttpPost("login-caregiver")]
+        [ProducesResponseType(typeof(AuthResponseDto), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<IActionResult> LoginCaregiver([FromBody] LoginDto request)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            try
+            {
+                var (result, token) = await _authService.LoginCaregiverAsync(
+                    request.Identifier,
+                    request.Password
+                );
+
+                if (!result.Success || string.IsNullOrEmpty(token))
+                {
+                    if (result.IsLockedOut)
+                    {
+                        return StatusCode(
+                            StatusCodes.Status403Forbidden,
+                            new { message = result.Message }
+                        );
+                    }
+                    return Unauthorized(new { message = result.Message });
+                }
+
+                var cookieOptions = _authService.GetJwtCookieOptions();
+                HttpContext.Response.Cookies.Append(CookieNames.Jwt, token, cookieOptions);
+
+                return Ok(
+                    new
+                    {
+                        message = result.Message,
+                        loggedInUser = $"{result.FirstName} {result.LastName}",
+                        roles = result.Roles,
+                    }
+                );
+            }
+            catch (JwtTokenGenerationException ex)
+            {
+                return StatusCode(500, new { message = ex.Message });
+            }
         }
 
         /// <summary>
