+++ HealthCareAB_v1/HealthCareAB_v1/Controllers/CaregiversController.cs
@@ -56,50 +56,4 @@ public class CaregiversController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
-
-    [HttpPost("login")]
-    [ProducesResponseType(typeof(AuthResponseDto), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    public async Task<IActionResult> Login([FromBody] LoginDto request)
-    {
-        if (!ModelState.IsValid)
-            return BadRequest(ModelState);
-        try
-        {
-            var (result, token) = await _authService.LoginCaregiverAsync(
-                request.Identifier,
-                request.Password
-            );
-
-            if (!result.Success || string.IsNullOrEmpty(token))
-            {
-                if (result.IsLockedOut)
-                {
-                    return StatusCode(
-                        StatusCodes.Status403Forbidden,
-                        new { message = result.Message }
-                    );
-                }
-                return Unauthorized(new { message = result.Message });
-            }
-
-            var cookieOptions = _authService.GetJwtCookieOptions();
-            HttpContext.Response.Cookies.Append(CookieNames.Jwt, token, cookieOptions);
-
-            return Ok(
-                new
-                {
-                    message = result.Message,
-                    loggedInUser = $"{result.FirstName} {result.LastName}",
-                    roles = result.Roles,
-                }
-            );
-        }
-        catch (JwtTokenGenerationException ex)
-        {
-            return StatusCode(500, new { message = ex.Message });
-        }
-    }
 }
