+++ HealthCareAB_v1/HealthCareAB_v1/Services/Interfaces/IAuthService.cs
@@ -5,7 +5,7 @@ namespace HealthCareAB_v1.Services.Interfaces
 {
     public interface IAuthService
     {
-        Task<AuthResponseDto> RegisterAsync(RegisterDto registerDto);
+        Task<AuthResponseDto> RegisterPatientAsync(RegisterDto registerDto);
 
         Task<(AuthResponseDto response, string? token)> LoginPatientAsync(
             string personalIdentityNumber,
